Minor Improvements
